Setup a new Dockerfile
This takes the minimum required to run a Laravel application, adds in
the dependencies required to run Pixelfed, plus some other helpful
additions.

* We need GD (`libgd-dev`), JPEGOptim (`jpegoptim`), OptiPNG
  (`optipng`), PNGQuant (`pngquant`) and `ffmpeg`
* The PHP GD extension, which needs `libfreetype-dev`,
  `libjpeg62-turbo-dev`, `libpng-dev`, `libxpm-dev`,
* The intl extension, which needs `libicu-dev`,
* The Redis extension which needs Redis from PECL,
* The Zip extension, which needs `libzip-dev`, `libmbedtls-dev`,
* Then PHP extensions (which don’t need additional dependencies):
  `bcmath`, `exit`, `iconv`, `pdo_mysql`, `mysqli` and `pcntl`

It’s recommended to set certain options in php.ini to match other
values, so we allow these to be overridden through environment
variables, but set sensible defaults.

Finally, we can’t cache the configuration because it’s provided after
the container is built but we can cache everything else.

https://docs.pixelfed.org/running-pixelfed/prerequisites.html
https://hub.docker.com/_/php
https://www.php.net/manual/en/configuration.file.php#example-36